+++ docs/scripts/config/index.js
@@ -4,7 +4,7 @@ import prismjs from 'prismjs';
 export const VERSION = version;
 export const isDev = process.env.NODE_ENV === 'development';
 export const isProd = process.env.NODE_ENV === 'production';
-export const domain = isProd ? '//material.balmjs.com' : '';
+export const domain = isProd ? '//v8.material.balmjs.com' : '';
 export const lazyLoadedTime = 1e3;
 
 // 同步 Sass variables