+++ docs/scripts/views/components/autocomplete.vue
@@ -62,7 +62,7 @@ const source = [
   },
   {
     name: 'BalmUI',
-    url: 'https://material.balmjs.com/'
+    url: 'https://v8.material.balmjs.com/'
   },
   {
     name: 'BalmScroll',