+++ CHANGELOG.md
@@ -1,6 +1,5 @@
-# [BalmUI](https://next-material.balmjs.com) ChangeLog
+# [BalmUI](https://material.balmjs.com) ChangeLog
 
-- `balm-ui@next` for Vue 3
 - [`balm-ui@8`](https://github.com/balmjs/balm-ui/tree/8.x) for Vue 2
 
 ## v9.40.3 / 2022-01-21